server/pyproject.toml
[project]
name = "file-upload-and-fixes"
version = "0.1.0"
description = ""
authors = [
    {name = "Ansh Shah"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.121.2,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.0,<3.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "python-multipart (>=0.0.18,<1.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.main]
init-hook = "import sys; import os; venv_path = os.path.join(os.path.dirname(os.path.abspath(__file__)), '.venv', 'lib', 'python3.12', 'site-packages'); sys.path.insert(0, venv_path) if os.path.exists(venv_path) else None"
